UdemyNew/requirements.txt
Django==2.2.9
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
